Split WarcraftClient.cs into separate files with partial classes
Split WarcraftClient.cs into separate files with partial classes, allowing the Community API methods to be separated.
